sanity check in publish.sh to prevent changes to published artefacts
Any line deleted or changed in releases/SHA1SUM is considered an error
because released artefacts must never change or be removed after they
are published.


Former-commit-id: 88ae61867b3e73d8638f6c74a1e5bb44b8040eec

+++ publish.sh
@@ -71,6 +71,16 @@ commit=$(git commit-tree $tree -p $parent -p $(git rev-parse HEAD) <<EOF
 Automatic commit of generated web site from $(git rev-parse HEAD)
 EOF
 )
+
+# Sanity check: did the new commit delete or change any lines in 
+# releases/SHA1SUM? This is verboten!
+deleted=$(git diff $parent..$commit --numstat -- releases/SHA1SUM | cut -f1)
+if [[ "$deleted" -gt 0 ]] ; then
+    echo "Checksum of released artefact changed!"
+    git diff $parent..$commit -- releases/SHA1SUM | cat
+    exit 1
+fi
+
 git update-ref refs/heads/published $commit $parent
 
 rm -rf "$D"
